Read the #186 unsigned-with-zero warnings; fix the two that hid bugs
Each #186-D is an unsigned value compared with zero. Read as a group:

- ptransfer.c:2327 (transfer_len >= 0): the 500-byte cap on a read reply
  only skipped building the header; the handler carried on. On the Temco
  private-Modbus path the CRC was then written at temp[transfer_len + 14],
  and with an invalid range nothing else had checked transfer_len, which
  is entitysize * count from the request - so a request could write two
  bytes up to ~64 KB past the 600-byte temp[]. An oversize read now gets no
  reply, as the other overrun checks in the handler already do, and the
  CRC write is bounded to temp[] as well.
- ptransfer.c:764 (invokeid_mstp >= 0): invokeid_mstp was a uint8_t
  holding the int that GetRemotePoint and Send_private_scan return, where
  a failed send is -1 or -7. It became 255 or 249, callers counted the
  send as made instead of bumping lose_count, and a stray reply with that
  id matched it. It is now an int, so the check means what it says.
- tstat_wifi.c:380 and 401, menuSet.c:83, modbus.c coil and serial-number
  ranges: dead lower bounds with a real upper bound beside them; removed
  with no change in behaviour.
- user_data.c:1485 (time_to_live < 0) is left as it is: remote panels have
  never expired, and turning expiry on would delete live panels on a real
  network. It goes on the README's To do instead.

Build: 0 errors, 465 warnings (7 fewer, none new).

+++ bacnet/private/user_data.h
@@ -288,7 +288,7 @@ typedef enum { WRITE_SD_OK = 0,WAIT_FOR_SD} E_OPERATE_SD;
 
 extern U32_T	timeCount1,timeCount2;
 //extern uint8_t invokeid_bip;
-extern uint8_t far invokeid_mstp;
+extern int far invokeid_mstp;
 extern uint8_t far flag_receive_netp;	// network points 
 extern uint8_t far flag_receive_netp_temcovar;
 //extern uint8_t far flag_receive_netp_temcoreg;
